requirements.txt
flask==2.3.3
openai==1.3.0
python-dotenv==1.0.0
langchain==0.0.267
langchain-community==0.0.16
langchain-openai==0.0.4
langchain-chroma==0.0.1
pandas==2.1.0
chromadb==0.4.17
PyPDF2==3.0.1
pillow==10.0.1
pytesseract==0.3.10
openpyxl==3.1.2
pyarrow==14.0.1
fastparquet==2023.10.1
werkzeug==2.3.7
tavily-python==0.2.1
markdown-it-py==3.0.0 